Tests/LibWeb/Text/expected/shared-array-buffer-atomics-wait-notify.txt

worker saw value: 42
